Add data_logging package
