+++ lsl.py
@@ -1,15 +1,12 @@
 """
 This module contains the LSL class, responsible for handling LSL input, collection, and formatting.
 """
-import csv
-
 import pylsl
-import pandas as pd
 import threading
 from datetime import datetime
 
 import config
-
+from data_logging.loggers import CollectedDataLogger
 
 class LSL:
     """
@@ -62,38 +59,21 @@ class LSL:
             LSL.collecting = False
             LSL.collection_thread.join()
             print("Data collection stopped. Saving collected data.")
-            LSL.__save_collected_data(config.COLLECTED_DATA_PATH)
+
+            if LSL.collected_data:
+                # Define column headers
+                eeg_channel_count = LSL.streams['EEG'].info().channel_count() if LSL.streams['EEG'] else 0
+                accelerometer_channel_count = LSL.streams['Accelerometer'].info().channel_count() if LSL.streams['Accelerometer'] else 0
+                fft_channel_count = LSL.streams['FFT'].info().channel_count() if LSL.streams['FFT'] else 0
+
+                CollectedDataLogger.record_data(LSL.collected_data,
+                                                eeg_channel_count,
+                                                accelerometer_channel_count,
+                                                fft_channel_count)
 
     #
     # HELPER METHODS
     #
-    @staticmethod
-    def __save_collected_data(path: str):
-        """
-        Function to save data collected after collection has been stopped.
-
-        :param path: The path to write the collected data to as a CSV file.
-        """
-        if LSL.collected_data:
-            # Define column headers
-            eeg_channel_count = LSL.streams['EEG'].info().channel_count() if LSL.streams['EEG'] else 0
-            accelerometer_channel_count = LSL.streams['Accelerometer'].info().channel_count() \
-                if LSL.streams['Accelerometer'] else 0
-            fft_channel_count = LSL.streams['FFT'].info().channel_count() if LSL.streams['FFT'] else 0
-
-            columns = ['Timestamp'] + \
-                      [f'EEG_{i + 1}' for i in range(eeg_channel_count)] + \
-                      [f'Accelerometer_{i + 1}' for i in range(accelerometer_channel_count)] + \
-                      [f'FFT_{i + 1}' for i in range(fft_channel_count)]
-
-            df = pd.DataFrame(LSL.collected_data, columns=columns)
-            df['Timestamp'] = pd.to_datetime(df['Timestamp'])
-            df = df.sort_values(by='Timestamp')
-            df.to_csv(path, index=False)
-            print("Collected data saved.")
-        else:
-            print("No data to save.")
-
     @staticmethod
     def __lsl_to_system_time(lsl_timestamp):  # TODO idk what type lsl_timestamp is
         """Convert an LSL timestamp to system time."""
@@ -137,50 +117,3 @@ class LSL:
                 flattened_data_row = [data_row['Timestamp']] + data_row['EEG'] + data_row['Accelerometer'] + data_row[
                     'FFT']
                 LSL.collected_data.append(flattened_data_row)
-
-
-class EventLogger:
-    """
-    A static class to log test events in a separate CSV
-    """
-    event_data = []
-
-    @staticmethod
-    def record_timestamp(event_name):
-        """
-        Record the time the data was collected along with the event
-
-        :param event_name: Event that was collected
-        """
-        timestamp = datetime.now()
-        EventLogger.event_data.append((event_name, timestamp))
-        print(f"{event_name}: {timestamp}")
-
-    @staticmethod
-    def save_to_csv() -> bool:
-        """
-        Write all data to the csv files
-
-        :return: If the writing was successful
-        """
-        if not EventLogger.event_data:
-            return False
-
-        # Wrap it to catch any file issues
-        try:
-            with open(config.EVENT_DATA_PATH, 'w', newline='') as file:
-                writer = csv.writer(file)
-                writer.writerow(["Event", "Timestamp"])
-                for data in EventLogger.event_data:
-                    writer.writerow(data)
-        except Exception:
-            return False
-
-        EventLogger.event_data.clear()
-        print(f"Event Data saved to {config.EVENT_DATA_PATH}")
-
-        return True
-
-    @staticmethod
-    def clear_data():
-        EventLogger.event_data.clear()
