+++ main.py
@@ -1,6 +1,7 @@
 import importlib
 import os
 
+from lsl import LSL
 from test_gui import TestGUI
 import config
 
@@ -44,4 +45,5 @@ class DataCollectorApp:
 
 if __name__ == '__main__':
     TestGUI.init_gui()
+    LSL.init_lsl_stream()
     DataCollectorApp.run()
